requirements.txt
# Mag 7 AI Sentiment Pipeline
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
vaderSentiment>=3.3.2
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
xgboost>=2.0.0
feedparser>=6.0.10
newsapi-python>=0.2.6
# Optional: LLM sentiment
# openai>=1.0.0
# transformers>=4.30.0
